Refactor iframe component for captcha and webauthn
